// src/app/models/interfaces/scanner-tip.interface.ts
export interface IScannerTip {
    Title: string;
    Information: string;
    InformationTitle: string;
    TopTextStart: string;
    TopTextProcessing: string;
    TopTextSuccess: string;
    TopTextError: string;
    BottomTextStart: string;
    BottomTextProcessing: string;
    BottomTextSuccess: string;
    BottomTextError: string;
}